sdk/tests/conformance/buffers/buffer-data-and-buffer-sub-data.html

<!--

/*
** Copyright (c) 2012 The Khronos Group Inc.
**
** Permission is hereby granted, free of charge, to any person obtaining a
** copy of this software and/or associated documentation files (the
** "Materials"), to deal in the Materials without restriction, including
** without limitation the rights to use, copy, modify, merge, publish,
** distribute, sublicense, and/or sell copies of the Materials, and to
** permit persons to whom the Materials are furnished to do so, subject to
** the following conditions:
**
** The above copyright notice and this permission notice shall be included
** in all copies or substantial portions of the Materials.
**
** THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
** CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
** MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
*/

-->

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../js/js-test-pre.js"></script>
<script src="../../js/webgl-test-utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<script>
"use strict";
description("Test bufferData/bufferSubData with or without ArrayBuffer input");

debug('Regression test for <a href="https://bugs.webkit.org/show_bug.cgi?id=41884">https://bugs.webkit.org/show_bug.cgi?id=41884</a> : <code>Implement bufferData and bufferSubData with ArrayBuffer as input</code>');

var wtu = WebGLTestUtils;
var gl = wtu.create3DContext();
var buf = null;
var array = null;
var arrayZero = null;
if (!gl) {
    testFailed("WebGL context does not exist");
} else {
    testPassed("WebGL context exists");

    bufferDataTest();
    bufferDataTestWithArrayBuffer();

    bufferSubDataTest();
}

function bufferDataTest() {
    debug("");
    debug("Test bufferData without ArrayBuffer input");

    buf = gl.createBuffer();
    shouldBeNonNull("buf");

    gl.bufferData(gl.ARRAY_BUFFER, 4, gl.STATIC_DRAW);
    wtu.glErrorShouldBe(gl, gl.INVALID_OPERATION, "no buffer bound");

    gl.bindBuffer(gl.ARRAY_BUFFER, buf);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR);

    gl.bufferData(gl.ARRAY_BUFFER, -4, gl.STATIC_DRAW);
    wtu.glErrorShouldBe(gl, gl.INVALID_VALUE,
        "calling bufferData when buffer size is negative should generate INVALID_VALUE");

    gl.bufferData(gl.ARRAY_BUFFER, 0, gl.STATIC_DRAW);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR,
        "calling bufferData when buffer size is zero should succeed");

    gl.bufferData(gl.ARRAY_BUFFER, 4, gl.STATIC_DRAW);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR, "calling bufferData should succeed");

    gl.bindBuffer(gl.ARRAY_BUFFER, null);
}

function bufferDataTestWithArrayBuffer() {
    debug("");
    debug("Test bufferData with ArrayBuffer input");

    array = new ArrayBuffer(128);
    shouldBeNonNull("array");

    buf = gl.createBuffer();
    shouldBeNonNull("buf");

    gl.bufferData(gl.ARRAY_BUFFER, array, gl.STATIC_DRAW);
    wtu.glErrorShouldBe(gl, gl.INVALID_OPERATION, "no buffer bound");

    gl.bindBuffer(gl.ARRAY_BUFFER, buf);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR);

    gl.bufferData(gl.ARRAY_BUFFER, null, gl.STATIC_DRAW);
    wtu.glErrorShouldBe(gl, gl.INVALID_VALUE,
        "calling bufferData when BufferDataSource is null should generate INVALID_VALUE");

    gl.bufferData(gl.ARRAY_BUFFER, array, gl.STATIC_DRAW);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR,
        "calling bufferData with ArrayBuffer should succeed");

    arrayZero = new ArrayBuffer(0);
    shouldBeNonNull("arrayZero");
    gl.bufferData(gl.ARRAY_BUFFER, arrayZero, gl.STATIC_DRAW);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR,
        "calling bufferData with zero-sized ArrayBuffer should succeed");

    gl.bindBuffer(gl.ARRAY_BUFFER, null);
}

function bufferSubDataTest() {
    debug("");
    debug("Test bufferSubData");

    buf = gl.createBuffer();
    gl.bindBuffer(gl.ARRAY_BUFFER, buf);
    gl.bufferData(gl.ARRAY_BUFFER, 128, gl.STATIC_DRAW);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR);

    array = new ArrayBuffer(64);

    gl.bufferSubData(gl.ARRAY_BUFFER, -10, array);
    wtu.glErrorShouldBe(gl, gl.INVALID_VALUE,
        "calling bufferSubData with ArrayBuffer when offset is negative should INVALID_VALUE");

    gl.bufferSubData(gl.ARRAY_BUFFER, 65, array);
    wtu.glErrorShouldBe(gl, gl.INVALID_VALUE, "buffer overflow");

    gl.bufferSubData(gl.ARRAY_BUFFER, -10, new Float32Array(8));
    wtu.glErrorShouldBe(gl, gl.INVALID_VALUE,
        "calling bufferSubData with ArrayBufferView when offset is negative should generate INVALID_VALUE");

    gl.bufferSubData(gl.ARRAY_BUFFER, 10, array);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR,
        "calling bufferSubData with ArrayBuffer should succeed");

    gl.bufferSubData(gl.ARRAY_BUFFER, 10, null);
    wtu.glErrorShouldBe(gl, gl.INVALID_VALUE,
        "calling bufferSubData when BufferDataSource is null should generate INVALID_VALUE");

    gl.bufferSubData(gl.ARRAY_BUFFER, 10, new Float32Array(0));
    wtu.glErrorShouldBe(gl, gl.NO_ERROR,
        "calling bufferSubData with zero-sized ArrayBufferView should succeed");

    gl.bindBuffer(gl.ARRAY_BUFFER, null);
}

var successfullyParsed = true;
</script>

<script src="../../js/js-test-post.js"></script>
</body>
</html>
